Added codechat model

See merge request Deltix/openai-apps/vertex-ai-to-openai-adapter!7

+++ universal_api/request.py
@@ -33,7 +33,7 @@ class CompletionParameters(BaseModel):
     n: Optional[int] = None
     stream: Optional[bool] = None
     top_p: Optional[float] = None
-    top_k: Optional[float] = None
+    top_k: Optional[int] = None
 
 
 # Direct translation of https://platform.openai.com/docs/api-reference/chat/create
